% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootciJMCS.R
\name{bootciJMCS}
\alias{bootciJMCS}
\title{Function for bootstrapped confidence interval}
\usage{
bootciJMCS(object, future_time)
}
\arguments{
\item{object}{fitted model}

\item{future_time}{time sequence at which estimates are required}
}
\value{
Returns bootstraped confidence interval for model fitted through FastJM
}
\description{
Function for bootstrapped confidence interval
}
