% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.jmbB}
\alias{predict.jmbB}
\title{predict.jmbB}
\usage{
\method{predict}{jmbB}(object, newdata, ...)
}
\arguments{
\item{object}{fitted model}

\item{newdata}{newdata}

\item{...}{others}
}
\value{
Survival Prediction for newdata from model fitted through \code{jmbB()}
}
\description{
predict.jmbB
}
