% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.jmcsB}
\alias{predict.jmcsB}
\title{predict.jmcsB}
\usage{
\method{predict}{jmcsB}(object, newdata, ...)
}
\arguments{
\item{object}{fitted model}

\item{newdata}{newdata}

\item{...}{others}
}
\value{
Survival Prediction for newdata from model fitted through
\code{jmcsB()}
}
\description{
predict.jmcsB
}
