% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.jmrmlB}
\alias{predict.jmrmlB}
\title{predict.jmrmlB}
\usage{
\method{predict}{jmrmlB}(object, newdata, ...)
}
\arguments{
\item{object}{fitted model}

\item{newdata}{newdata}

\item{...}{others}
}
\value{
Survival Prediction for newdata from model fitted through \code{jmrmlB()}
}
\description{
predict.jmrmlB
}
