% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.jmstB}
\alias{predict.jmstB}
\title{predict.jmstB}
\usage{
\method{predict}{jmstB}(object, newdata, ...)
}
\arguments{
\item{object}{fitted model}

\item{newdata}{newdata}

\item{...}{others}
}
\value{
Survival Prediction for newdata from model fitted through
\code{jmstB()}
}
\description{
predict.jmstB
}
