% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.jmbB}
\alias{print.jmbB}
\title{print.}
\usage{
\method{print}{jmbB}(x, ...)
}
\arguments{
\item{x}{fitted object}

\item{...}{others}
}
\value{
prints table containing various parameter estimates,
        SE, P- value for both survival and longitudinal submodel,
        if the model is bayesian it includes their credible interval too.
}
\description{
print method for class 'jmbB'
}
